Remove changes from deepseek_v3 folder
