+++ models/demos/deepseek_v3_d_p/utils/transformer_helpers.py
@@ -494,14 +494,14 @@ def load_and_compute_layer_by_layer(
         LayerByLayerResult(state_dict=None, ref_snapshots, ref_kvpe_list)
         Note: state_dict is always None (cache built to disk instead)
     """
-    from models.demos.deepseek_v3.utils.config_helpers import detect_language_model_prefix, sub_state_dict
+    from models.demos.deepseek_v3.utils.config_helpers import sub_state_dict
     from models.demos.deepseek_v3.utils.lazy_state_dict import LazyStateDict
-    from models.demos.deepseek_v3.utils.test_utils import dequantize_state_dict
     from models.demos.deepseek_v3_d_p.tt.moe.tt_moe_gate_prefill import GateComputeMode
     from models.demos.deepseek_v3_d_p.tt.tt_distributed_rms_norm import TtDistributedRmsNorm
     from models.demos.deepseek_v3_d_p.tt.tt_lm_head import TtLMHead
     from models.demos.deepseek_v3_d_p.tt.tt_parallel_embedding import TtParallelEmbedding
     from models.demos.deepseek_v3_d_p.tt.tt_prefill_block import TtPrefillBlock
+    from models.demos.deepseek_v3_d_p.utils.test_utils import dequantize_state_dict, detect_language_model_prefix
 
     if gate_fallback_mode is None:
         gate_fallback_mode = GateComputeMode.HOST_ALL
